Update dependencies in go.mod and go.sum
This commit has updated the version of the dependencies in go.mod and go.sum. Several versions of dependencies including but not limited to 'gopenpgp/v2', 'go-msgauth', and 'go-mail' were updated to newer versions. This ensures the project is keeping up-to-date with the latest releases and improvements from these dependencies.